Add a generic exception handler to cbmc-viewer.
We wrap the main function of cbmc-viewer with a try/except that prints
a stack trace along with instructions to the user on how to submit an
issue to github if the trace does not help them debug the problem
themselves.

This commit also adds a bit of document to the top of the python script
that implements the cbmc-viewer regression testing.