Improve parsing of "make goto" output, remove symbolic links in srcdir.
The output of "make goto" may break a single command over many lines, so
do a better job of parsing the output for FILE in "goto-cc ... -o FILE".
Also, the preprocessor may remove symbolic links from filenames it uses
in the preprocessor output, so the files listed may not begin with srcdir
if srcdir itself involves symbolic links.