Add concurrency, configuration, and cbmc warning reporting.
Add a warning section to the report for warnings issued by cbmc.
Report on functions with missing function bodies, and report
separately on other warnings (like ignoring byte-enable...).

Enable concurrent invocation of cbmc-viewer.  Use the process id to
generate a unique name for the TAGS file so concurrent invocations of
the script will not try to use the same TAGS file.  The right solution
is to use the tempfile module to get unique files in a temporary
directory, but this will not work because etags writes paths to the
TAGS file that are relative from the TAGS file to the source file.

Add a --config flag to specify a json file (default cbmc-viewer.py)
giving viewer configuration information.  Ultimately, all command line
arguments could be specified here.  For now, the only entry is
'expected-missing-functions' which is a list of names of function we
expect to have a missing function body.  If this file can be read, the
list of missing functions above is split into "expected" and
"unexpected" missing functions.