Make viewer accept multiple ';' blocks in coverage descriptions.
It is possible for the source code lines in a single basic block
to come from multiple files.  The classical example was function
inlining, but cbmc stopped doing inlining, so this problem went
away.  It is still possible when source code appears in nested
header files.  This patch adds back to viewer the ability to
parse the basic block desciptions with lines from multiple files.