Remove old Ubuntu16/Ubuntu18 builds.
This removes the old Ubuntu16 and Ubuntu18 work and relies now on the
original Debian work to build the Ubuntu packages.  This implements a
new API for that work to create cbmc and cbmc-latest packages from
tarballs and from a clone of the repository.  This also adds as a
comment the workflow trigger to use when this is merged into the cbmc
repo.